merge sprint ending 10/3/23

+++ packages/react-app-revamp/supabase_schemas/analytics_contest_participants_v3 .sql
@@ -5,8 +5,8 @@ create table
     user_address character varying not null,
     network_name character varying not null,
     proposal_id character varying null,
-    vote_amount integer null,
-    created_at integer null,
+    vote_amount numeric null,
+    created_at int4 null,
     constraint analytics_contest_participants_v3_pkey primary key (uuid)
   ) tablespace pg_default;
